--- ha-business/src/main/java/jp/co/ha/business/api/aws/AwsS3Key.java
@@ -1,45 +1,0 @@
-package jp.co.ha.business.api.aws;
-
-import jp.co.ha.common.type.BaseEnum;
-
-/**
- * AWSのS3のキー列挙
- *
- * @version 1.0.0
- */
-public enum AwsS3Key implements BaseEnum {
-
-    /** 健康情報一括登録CSVファイル配置キー */
-    HEALTHINFO_FILE_REGIST("healthinfo-file-regist/"),
-    /** 健康情報照会CSVファイルの配置キー */
-    HEALTHINFO_FILE_REFERENCE("healthinfo-file-reference/"),
-    /** お知らせ一覧JSONファイルの配置キー */
-    NEWS_JSON("news/"),
-    /** 月次健康情報集計CSV配置キー */
-    MONTHLY_HEALTHINFO_SUMMARY("monthly/healthinfo/"),
-    /** Slack接続情報キー */
-    SLACK_CONNECTION_DATA("slack/slack_connection_data.json"),
-    /** メモ情報配置キー */
-    NOTE_FILE("note/"),
-    /** 問い合わせキー */
-    INQUIRY("inquiry/");
-
-    /**
-     * コンストラクタ
-     *
-     * @param value
-     *     値
-     */
-    private AwsS3Key(String value) {
-        this.value = value;
-    }
-
-    /** 値 */
-    private String value;
-
-    @Override
-    public String getValue() {
-        return this.value;
-    }
-
-}
