+++ ha-business/src/main/java/jp/co/ha/business/api/node/type/NodeApiType.java
@@ -54,7 +54,7 @@ public enum NodeApiType implements BaseEnum {
     }
 
     /**
-     * apiNameTypeを返す
+     * API名を返す
      *
      * @return apiNameType
      */
