+++ ha-business/src/main/java/jp/co/ha/business/api/root/type/RootApiType.java
@@ -40,7 +40,7 @@ public enum RootApiType implements BaseEnum {
     }
 
     /**
-     * apiNameTypeを返す
+     * API名を返す
      *
      * @return apiNameType
      */
