+++ ha-root/src/main/java/jp/co/ha/root/contents/healthinfo/controller/HealthInfoEditApiController.java
@@ -74,6 +74,7 @@ public class HealthInfoEditApiController extends
      * @throws BaseException
      *     API呼び出しに失敗した場合
      */
+    @SuppressWarnings("deprecation")
     @PutMapping(value = "healthinfo/{seq_health_info_id}", produces = {
             MediaType.APPLICATION_JSON_VALUE })
     public ResponseEntity<HealthInfoEditApiResponse> edit(
@@ -105,7 +106,6 @@ public class HealthInfoEditApiController extends
         } else {
 
             // トークン発行API実施
-            @SuppressWarnings("deprecation")
             TokenApiResponse tokenResponse = tokenApiComponent.callTokenApi(
                     request.getSeqUserId(), transactionId);
